requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.5.0
astroNN>=1.1.0
torch-fidelity>=0.3.0
tensorboard>=2.13.0
h5py>=3.8.0
scipy>=1.10.0
tqdm>=4.65.0
